+++ app/controllers/application_controller.rb
@@ -3,7 +3,6 @@
 # This controller also handles several static pages in views/application
 class ApplicationController < ActionController::Base
   before_action :configure_permitted_parameters, if: :devise_controller?
-  before_action :authenticate_user!, only: [:dashboard]
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
@@ -21,8 +20,7 @@ class ApplicationController < ActionController::Base
     (params_column, params_direction) = params_order.to_s.strip.downcase.split(' ')
     direction = (params_direction == 'desc' ? 'DESC' : nil)
     column_name = (model.column_names.collect { |c| model.table_name + '.' + c }.find { |c| c == params_column })
-    order = column_name.blank? ? default_order : [column_name, direction].compact.join(' ')
-    order
+    column_name.blank? ? default_order : [column_name, direction].compact.join(' ')
   end
 
   def check_editor